+++ CHANGELOG.md
@@ -19,6 +19,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Scale.getIndex
 
+- Scale.degreeInOctave
+
 
 ## 2.6.0 (2023-11-10)